ai-bi-chatbot-project-main/Team5_module/backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.1
python-dotenv==1.0.0
Werkzeug==2.3.6
requests==2.31.0
google-generativeai==0.3.0
ReportLab==4.0.4
WeasyPrint==58.1
Pillow==10.0.0
pytest==8.3.3
